setup/requirements.txt
cdo==1.5.6
cdsapi==0.5.1
motuclient==1.8.4
netCDF4==1.6.2
numpy==1.22.2
pandas==1.5.0
psycopg2==2.9.5
python-dotenv==0.21.1
xarray==0.20.1
